client property segmentation
